requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.13.0
argcomplete==3.6.3
asciistuff==1.3.3
astroid==4.0.4
attrs==25.4.0
audioop-lts==0.2.2
backoff==2.2.1
backports.pbkdf2==0.1
beautifulsoup4==4.14.3
blinker==1.9.0
branca==0.8.2
brotli==1.2.0
bs4==0.0.2
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cinemagoer==2023.5.1
click==8.3.1
codext==1.15.11
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.8
cowpy==1.1.5
cryptography @ file:///home/builder/.termux-build/python-cryptography/src
cssselect2==0.9.0
dateparser==1.3.0
dicttoxml==1.7.16
dill==0.4.1
distlib==0.4.0
distro==1.9.0
engineering-notation==0.12.0
filelock==3.25.2
find-github-email==1.2.6
flake8==7.3.0
Flask==3.1.3
flask-cors==6.0.2
folium==0.20.0
fonttools==4.62.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google==3.0.0
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
googleapis-common-protos==1.73.0
googlesearch-python==1.3.0
greenlet==3.3.2
groq==0.15.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
humanfriendly==10.0
idna==3.11
IMDbPY==2022.7.9
isort==8.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
json2html==1.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
legacy-cgi==2.6.4
legacycrypt==0.3
lxml @ file:///home/builder/.termux-build/python-lxml/src
markdown-it-py==4.0.0
markdown2==2.5.5
MarkupSafe==3.0.3
maturin==1.12.6
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
narwhals==2.18.1
netaddr==1.3.0
netifaces==0.11.0
numpy @ file:///home/builder/.termux-build/python-numpy/src/dist/numpy-2.4.3-cp313-cp313-android_aarch64.whl#sha256=8ef35b7e7291aa8501ff0be8f354f3b089c4e0c08e9a905301b3551814601263
openai==2.30.0
opencage==3.3.0
packaging==26.0
pandas @ file:///home/builder/.termux-build/python-pandas/src/dist/pandas-3.0.2-cp313-cp313-android_aarch64.whl#sha256=1fceb33c6324048bc6ee6bee30928f882d664f2a80e0f0f984b541f38bfbbd27
patchy==2.10.0
pathlib2==2.3.7.post1
pbkdf2==1.3
peewee==4.0.2
phonenumbers==9.0.26
piexif==1.1.3
pillow @ file:///home/builder/.termux-build/python-pillow/src
ping3==5.1.5
platformdirs==4.9.4
plyer==2.1.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
pyarrow @ file:///home/builder/.termux-build/libarrow-cpp/src/python/dist/pyarrow-23.0.1-cp313-cp313-android_aarch64.whl#sha256=68feb8d1d53d02724d25e089b5a5d0af321aa5e459e764fd31e01e1407767601
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
pydyf==0.12.1
pyfiglet==1.0.4
pyflakes==3.4.0
Pygments==2.19.2
pylint==4.0.5
pyminizip==0.2.6
pynput==1.8.1
pypandoc==1.17
pyparsing==3.3.2
pypdf==6.9.2
pyphen==0.17.2
pyTelegramBotAPI==4.32.0
python-apt @ file:///home/builder/.termux-build/python-apt/src
python-dateutil==2.9.0.post0
python-discovery==1.1.3
python-dotenv==1.2.2
python-http-client==3.3.7
python-magic==0.4.27
python-slugify==8.0.4
python-telegram-bot==21.10
pytz==2026.1.post1
PyYAML==6.0.3
qrcode==8.2
referencing==0.37.0
regex==2026.2.28
requests==2.33.1
resolver==0.2.1
rich==14.3.3
rpds-py==0.30.0
scapy==2.7.0
scipy @ file:///home/builder/.termux-build/python-scipy/src/dist/scipy-1.17.1-cp313-cp313-android_aarch64.whl#sha256=983a3b6ea280209902b8cc8c5c47d881237522224142999ddf2c8e084ffb2db3
sendgrid==6.12.5
setuptools==82.0.1
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
speedtest-cli==2.1.3
sploitkit==0.6.0
SQLAlchemy==2.0.48
streamlit==1.56.0
stringify==0.1.1
stripe==14.4.1
telnetlib3==4.0.1
tenacity==9.1.4
terminaltables==3.1.10
text-unidecode==1.3
tinycss2==1.5.1
tinyhtml5==2.1.0
tinyscript==1.31.2
tk-tools==0.17.0
toml==0.10.2
tomlkit==0.14.0
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vininfo==1.9.2
virtualenv==21.2.0
watchdog==6.0.0
wcwidth==0.6.0
weasyprint==68.1
webencodings==0.5.1
Werkzeug==3.1.6
wheel==0.46.3
wifi==0.3.8
xmltodict==1.0.4
xyzservices==2025.11.0
yarl==1.23.0
zopfli==0.4.1
